added async code highlight components
